Ver_0.0.0/requirements.txt
paddlepaddle==3.1.0
paddleocr==3.1.0
transformers>=4.30.0
torch>=2.6.0
pyqt6==6.9.1
numpy==1.26.4
Pillow==10.4.0
mss==9.0.0
requests==2.32.3
google-generativeai==0.8.3